stop metric change process when clicking elsewhere
Previously, to stop metric change process, i.e. to close the dropdown
menu, user had to click on the dropdown menu somewhere. Even in the case
the user did not want to change metric, the user had to either click the
header or the same metric as was already chosen. This change makes it so
that clicking anywhere besides the dropdown menu will quit the metric
change process.