bugfix: don't move widget with topbar
Previously, the widget would move depending on if the topbar was visible
or not. If the topbar was visible, the widget was beneath it. However,
if topbar was hidden, widget would move up. Problem with this approach
is that as soon as the user moves the mouse to click on the widget, the
topbar appears and the widget is moved down. The result is that the
widget can never be clicked.

Fix by having the widget statically positioned under the topbar
regardless if topbar is hidden or not.

Note that if the topbar widget is missing altogether, spectator hud is
placed at the top of the screen.