fix incorrect use of color in drawVSModeMetrics()
To use white color for text, we used the command gl.Color(). However, to
set text color, one has to use SetTextColor rather than Color(). This
meant that the text color could randomly be something different.

Meanwhile, in the function drawVSBar() there was a missing call to
gl.Color() with white color. This worked previously since drawVSBar() is
a subfunction to drawVSModeMetrics(). But with the fix mentioned above,
it no longer worked. Therefore, we need to fix drawVSBar() by adding a
call to gl.Color().