add unit cache
Add a unit cache that is used for army value and build power
calculations. With the cache, the widget no longer has to loop through
all units over and over to keep track of current statistics. Instead,
upon initialization, the widget once loops through all units to populate
the cache. Afterwards, the widget only updates as units are created,
destroyed and gifted.

Issue #13