performance: preallocate table vsModeStats
The object vsModeStats is used to store statistics for each team.
Previously, this object was released and allocated over and over as the
data got outdated and had to be updated. With this change that data is
still constantly updated, but the memory used is recycled rather than
having the whole object recreated from scratch.

This change is observed to drop memory allocations/deallocation in 8v8
games from around 10 MB/s down to 6 MB/s.