playernames: cache names on first game frame
Previously, we called GetPlayerInfo() over and over again in order to
get player names. Instead, we can cache player names on the first game
frame and reuse the names from the cache.

This is preliminary work for collecting player start positions in order
to allow sorting vsmode statistics according to spawn positions.