optimization: reorganize vsModeStats
Previously, we organized data in vsModeStats so that under each allyTeam
we would have teamID's and then metrics. However, we always access
teamID's and metrics in the opposite order. Metric is in the outer loop
and teamID's in the inner. Therefore, reorganize the data accordingly.

The optimization resulted in no measurable performance improvement.